third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement.getCTM-empty-viewBox-expected.txt

This is a testharness.js-based test.
[FAIL] SVGGraphicsElement.getCTM and co with empty viewbox
  assert_equals: getScreenCTM: e expected 0 but got 8
Harness: the test ran to completion.
